[Refactor] Inline DFlash2 grouped convolution
